requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.77
botocore==1.29.77
certifi==2022.12.7
charset-normalizer==3.0.1
dateparser==1.1.7
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
multidict==6.0.4
pycryptodome==3.17
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
websockets==10.4
yarl==1.8.2
